+++ Arma3ObjectBuilder/ui/__init__.py
@@ -13,4 +13,3 @@ from . import tool_proxies
 from . import tool_rigging
 from . import tool_scripts
 from . import tool_utilities
-from . import tool_validation
